+++ jkube-kit/config/resource/src/main/java/org/eclipse/jkube/kit/config/resource/ServiceAccountConfig.java
@@ -28,6 +28,8 @@ public class ServiceAccountConfig {
 
   private String name;
   private String deploymentRef;
+  private Boolean bindToAllControllers;
+  private Boolean generate;
 
 }